+++ src/lib/bridge/handlers/execute_script.ts
@@ -1,6 +1,6 @@
 import {EmptyResponseHandler} from './main';
 import {RequestType} from './types';
-import {PrivilegedHandler} from './wrappers';
+import {PrivilegedHandler} from '../wrappers/privileged';
 
 interface ExecuteScriptRequest {
     path: string;
